Refactor tortoise_api cog
